addtoplaylistbutton is now modular
