+++ src/Components/AddSongToPlaylist/AddSongToPlaylist.jsx
@@ -1,12 +1,12 @@
-const AddSongToPlaylist = () => {
+// const AddSongToPlaylist = () => {
 
 
 
-    return (
-        <div>
-            <button>Add to playlist</button>
-        </div>
-    )
-}
+//     return (
+//         <div>
+//             <button>Add to playlist</button>
+//         </div>
+//     )
+// }
 
-export default AddSongToPlaylist
+// export default AddSongToPlaylist
