+++ src/Components/PlaylistDisplay/PlaylistDisplay.jsx
@@ -1,14 +1,12 @@
-// PlaylistDisplay.jsx
 import React from "react";
 import { useLocation } from "react-router-dom";
 import useFetchTracks from "../Song/GetTrackUrl";
-// import usePlaylistFetch from "./hooks/usePlaylistFetch";
-// import TrackList from "./Components/TrackList";
-// import PlaylistHeader from "./Components/PlaylistHeader";
-import "./playlistDisplay.css";
 import usePlaylistFetch from "../../Hooks/usePlaylistFetch";
 import TrackList from "../TrackList";
-import PlaylistHeader from "../PlaylistHeader";
+import PlaylistHeader from "./PlaylistHeader";
+import "./playlistDisplay.css";
+
+
 const PlaylistDisplay = () => {
   const location = useLocation();
   const { playlist, trackUrl } = location.state || {};
