--- src/Components/Song/playlistMenu.css
@@ -1,10 +1,0 @@
-.playlist-sub-container{
-    display: flex;
-    flex-direction: row;
-    width:100%;
-    justify-content: flex-start;
-    align-content: center;
-    gap:1.5rem;
-    padding:1rem;
-    background-color: rgba(0, 0, 0, 0.093);
-}
